Move Physics system into Shared.
