+++ SS14.Shared/Interfaces/Physics/ICollidable.cs
@@ -1,7 +1,7 @@
 using SFML.Graphics;
 using SS14.Shared.Interfaces.GameObjects;
 
-namespace SS14.Client.Interfaces.Collision
+namespace SS14.Shared.Interfaces.Physics
 {
     public interface ICollidable
     {
