+++ SS14.Shared/Physics/CollidablePoint.cs
@@ -1,0 +1,29 @@
+﻿using SFML.System;
+
+namespace SS14.Shared.Physics
+{
+    internal enum CollidablePointIndex
+    {
+        TopLeft,
+        TopRight,
+        BottomRight,
+        BottomLeft
+    }
+
+    /// <summary>
+    /// This represents a point of a collision AABB
+    /// </summary>
+    internal struct CollidablePoint
+    {
+        public Vector2f Coordinates;
+        public CollidablePointIndex Index;
+        public CollidableAABB ParentAABB;
+
+        public CollidablePoint(CollidablePointIndex index, Vector2f coordinates, CollidableAABB parentAABB)
+        {
+            Index = index;
+            Coordinates = coordinates;
+            ParentAABB = parentAABB;
+        }
+    }
+}
